bot/pyproject.toml
[project]
name = "tania-signal-copier"
version = "0.1.0"
description = "Telegram to MetaTrader 5 Signal Bot - Automates forex trading by reading signals from Telegram and executing on MT5"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "telethon>=1.37.0",
    "PyQt6>=6.7.0",
    "groq>=0.19.0",
    "cerebras-cloud-sdk>=1.0.0",
    "python-dotenv>=1.0.1",
    "numpy>=2.2.0",
    "MetaTrader5>=5.0.0; sys_platform == 'win32'",
    "siliconmetatrader5>=1.1.0; sys_platform == 'darwin'",
    "rpyc==5.0.1; sys_platform == 'linux'",
    "plumbum>=1.7.0; sys_platform == 'linux'",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pyright>=1.1.390",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ============== Ruff Configuration ==============
[tool.ruff]
target-version = "py313"
line-length = 100
exclude = [
    ".venv",
    "__pycache__",
    ".git",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort (import sorting)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "PERF",   # Perflint (performance)
    "RUF",    # Ruff-specific rules
    "ASYNC",  # flake8-async
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ============== Pyright Configuration ==============
[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "standard"
reportUnusedImport = "error"
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"

# ============== Pytest Configuration ==============
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --cov=src --cov-report=term-missing"
markers = [
    "integration: mark test as integration test requiring MT5 Docker container",
    "slow: mark test as slow running",
]

# ============== Coverage Configuration ==============
[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
